fix consentQuestions: redirection links added
